pip_requirements.txt
python-pygame
-e https://github.com/TripleSpeeder/LTWheelConf